backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Image processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
pytesseract==0.3.13
ultralytics==8.1.0
imagehash==4.3.1

# PDF generation
weasyprint==59.0
pydyf==0.10.0
jinja2==3.1.3

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
websocket-client==1.7.0
